Removed legacy useGetCurrentView hook
